refactor(client/go): gather the Composer system under one composer/ root
composer/ (core), composer/meta (config-metadata side table, formerly
outline/connmeta — nothing in it was connection- or Outline-specific),
and composer/netconfig (strategy configs) now form one SDK-bound,
app-agnostic package tree, following the stdlib go/ family pattern.
Outline policy stays under client/go/outline/. DESIGN.md D14 records
the layout rule and the audience map.
