+++ client/go/composer/netconfig/block_test.go
