+++ client/go/composer/netconfig/dialendpoint.go
