+++ client/go/composer/netconfig/dialendpoint_test.go
