+++ client/go/composer/netconfig/interfaces.go
