+++ client/go/composer/netconfig/websocket.go
