+++ client/go/composer/netconfig/websocket_test.go
